frontend reorganization
