+++ UI/map-me/src/components/footer/footer.tsx
