+++ UI/map-me/src/components/landingPage/featureSection.tsx
